+++ src/datasources/asset/AssetDataSource.ts
@@ -12,6 +12,7 @@ import {
   AssetDataSourceOptions,
   AssetQuery,
   AssetQueryType,
+  AssetQueryReturnType
 } from './types/types';
 import { CalibrationForecastDataSource } from './data-sources/calibration-forecast/CalibrationForecastDataSource';
 import { AssetSummaryQuery } from './types/AssetSummaryQuery.types';
@@ -20,14 +21,16 @@ import { ListAssetsQuery } from './types/ListAssets.types';
 import { ListAssetsDataSource } from './data-sources/list-assets/ListAssetsDataSource';
 import { AssetSummaryDataSource } from './data-sources/asset-summary/AssetSummaryDataSource';
 import { AssetModel, AssetsResponse } from 'datasources/asset-common/types';
-import { QUERY_LIMIT } from './constants/constants';
 import { transformComputedFieldsQuery } from 'core/query-builder.utils';
 import { AssetVariableQuery } from './types/AssetVariableQuery.types';
+import { defaultListAssetsVariable } from './defaults';
+import { TAKE_LIMIT } from './constants/ListAssets.constants';
 
 export class AssetDataSource extends DataSourceBase<AssetQuery, AssetDataSourceOptions> {
   private assetSummaryDataSource: AssetSummaryDataSource;
   private calibrationForecastDataSource: CalibrationForecastDataSource;
   private listAssetsDataSource: ListAssetsDataSource;
+  private assetQueryReturnType: AssetQueryReturnType = AssetQueryReturnType.AssetIdentification;
 
   constructor(
     readonly instanceSettings: DataSourceInstanceSettings<AssetDataSourceOptions>,
@@ -89,25 +92,51 @@ export class AssetDataSource extends DataSourceBase<AssetQuery, AssetDataSourceO
     return this.listAssetsDataSource;
   }
 
+  getQueryReturnType(): AssetQueryReturnType {
+    return this.assetQueryReturnType;
+  }
+
+  setQueryReturnType(queryReturnType: AssetQueryReturnType): void {
+    this.assetQueryReturnType = queryReturnType;
+  }
+
   async metricFindQuery(query: AssetVariableQuery, options: LegacyMetricFindQueryOptions): Promise<MetricFindValue[]> {
+    let listAssetsTake = this.patchListAssetQueryVariable(query).take;
+    if (!this.isTakeValid(listAssetsTake)) {
+      return [];
+    }
     let assetFilter = query?.filter ?? '';
     assetFilter = transformComputedFieldsQuery(
       this.templateSrv.replace(assetFilter, options.scopedVars),
       this.listAssetsDataSource.assetComputedDataFields,
       this.listAssetsDataSource.queryTransformationOptions
     );
-    const assetsResponse: AssetsResponse = await this.listAssetsDataSource.queryAssets(assetFilter, QUERY_LIMIT);
-    return assetsResponse.assets.map(this.getAssetNameForMetricQuery);
+    const assetsResponse: AssetsResponse = await this.listAssetsDataSource.queryAssets(assetFilter, listAssetsTake);
+    return assetsResponse.assets.map((asset) => this.getAssetNameForMetricQuery(asset));
   }
 
   private getAssetNameForMetricQuery(asset: AssetModel): MetricFindValue {
     const vendor = asset.vendorName ? asset.vendorName : asset.vendorNumber;
     const model = asset.modelName ? asset.modelName : asset.modelNumber;
     const serial = asset.serialNumber;
+    let assetValue: string;
 
     const assetName = !asset.name ? `${serial}` : `${asset.name} (${serial})`;
-    const assetValue = `Assets.${vendor}.${model}.${serial}`
+    
+    if (this.assetQueryReturnType === AssetQueryReturnType.AssetId) {
+      assetValue = asset.id;
+    } else {
+      assetValue = `Assets.${vendor}.${model}.${serial}`;
+    }
 
     return { text: assetName, value: assetValue };
+  }
+
+  public patchListAssetQueryVariable(query: AssetQuery): AssetVariableQuery {
+    return { ...defaultListAssetsVariable, ...query } as AssetVariableQuery;
+  }
+
+  private isTakeValid(take?: number): boolean {
+    return take !== undefined && take >= 0 && take <= TAKE_LIMIT;
   }
 }