+++ src/datasources/asset/defaults.ts
@@ -1,4 +1,5 @@
-import { OutputType } from "./types/ListAssets.types"
+import { QUERY_LIMIT } from "./constants/constants"
+import { AssetFilterPropertiesOption, OutputType } from "./types/ListAssets.types"
 
 export const defaultAssetSummaryQuery = {
 }
@@ -11,5 +12,18 @@ export const defaultCalibrationForecastQuery = {
 export const defaultListAssetsQuery = {
     filter: "",
     outputType: OutputType.Properties,
-    take: 1000,
+    take: QUERY_LIMIT,
+    properties: [AssetFilterPropertiesOption.VendorName, AssetFilterPropertiesOption.AssetName, AssetFilterPropertiesOption.ModelName, AssetFilterPropertiesOption.Workspace, AssetFilterPropertiesOption.Location],
+}
+
+export const defaultListAssetsQueryForOldPannels = {
+    filter: "",
+    outputType: OutputType.Properties,
+    take: QUERY_LIMIT,
+    properties: Object.values(AssetFilterPropertiesOption),
+}
+
+export const defaultListAssetsVariable = {
+    filter: "",
+    take: QUERY_LIMIT,
 }